feat(user): handle delete user

+++ src/controllers/UserController.ts
@@ -1,8 +1,8 @@
-import { Request, Response } from 'express'
-import { validate } from 'class-validator'
+import {Request, Response} from 'express'
+import {validate} from 'class-validator'
 
-import { User } from '../entity/User'
-import { AppDataSource } from '../data-source'
+import {User} from '../entity/User'
+import {AppDataSource} from '../data-source'
 import Utils from '../utils'
 
 class UserController {
@@ -31,7 +31,7 @@ class UserController {
             res.status(200).json({
                 'data': users
             })
-        } else{
+        } else {
             res.status(400).json({
                 'message': 'API version does not match.'
             })
@@ -60,7 +60,7 @@ class UserController {
                     'message': 'User not found'
                 })
             }
-        } else{
+        } else {
             res.status(400).json({
                 'message': 'API version does not match.'
             })
@@ -69,7 +69,7 @@ class UserController {
 
     static newUser = async (req: Request, res: Response) => {
         //Get parameters from the body
-        const { username, password } = req.body
+        const {username, password} = req.body
         const user = new User()
         user.username = username
         user.password = password
@@ -102,7 +102,7 @@ class UserController {
         const id = req.body.id
 
         //Get values from the body
-        const { username, role } = req.body
+        const {username, role} = req.body
 
         //Try to find user on database
         const userRepository = AppDataSource.getRepository(User)
@@ -138,20 +138,36 @@ class UserController {
     }
 
     static deleteUser = async (req: Request, res: Response) => {
-        //Get the ID from the url
-        const id = req.body.id
+        const version = Utils.getApiVersion(req.baseUrl, res)
 
-        const userRepository = AppDataSource.getRepository(User)
-        try {
-            await userRepository.findOneOrFail(id)
-        } catch (error) {
-            res.status(404).send('User not found')
-            return
+        if (version === 'v1') {
+            //Get the ID from the url
+            const id = Number(req.body.id)
+
+            const userRepository = AppDataSource.getRepository(User)
+            try {
+                await userRepository.findOneOrFail({
+                    where: {
+                        id: id
+                    }
+                });
+            } catch (error) {
+                res.status(404).json({
+                    'message': 'User not found'
+                })
+                return
+            }
+            await userRepository.delete(id)
+
+            //After all send a 204 (no content, but accepted) response
+            res.status(200).json({
+                'message': 'User deleted'
+            })
+        } else {
+            res.status(400).json({
+                'message': 'API version does not match.'
+            })
         }
-        await userRepository.delete(id)
-
-        //After all send a 204 (no content, but accepted) response
-        res.status(204).send()
     }
 }
